+++ HPXMLtoOpenStudio/resources/hvac_sizing.rb
@@ -75,6 +75,7 @@ class HVACSizing
 
       hvac_sizing_values = HVACSizingValues.new
       apply_hvac_loads(hvac, hvac_sizing_values, bldg_design_loads)
+      apply_hvac_heat_pump_logic(hvac_sizing_values, hvac)
       apply_hvac_equipment_adjustments(hvac_sizing_values, weather, hvac)
       apply_hvac_installation_quality(hvac_sizing_values, weather, hvac)
       apply_hvac_fixed_capacities(hvac_sizing_values, hvac)
@@ -1134,6 +1135,27 @@ class HVACSizing
     hvac_sizing_values.Cool_Load_Lat = bldg_design_loads.Cool_Lat * hvac.CoolingLoadFraction
   end
 
+  def self.apply_hvac_heat_pump_logic(hvac_sizing_values, hvac)
+    # If HERS/MaxLoad methodology, uses at least the larger of heating and cooling loads for heat pump sizing (required for ERI).
+    if [HPXML::HVACTypeHeatPumpAirToAir,
+        HPXML::HVACTypeHeatPumpMiniSplit,
+        HPXML::HVACTypeHeatPumpGroundToAir,
+        HPXML::HVACTypeHeatPumpWaterLoopToAir,
+        HPXML::HVACTypeHeatPumpPTHP].include? hvac.CoolType
+      if (@hpxml.header.heat_pump_sizing_methodology != HPXML::HeatPumpSizingACCA) && (hvac.CoolingLoadFraction > 0) && (hvac.HeatingLoadFraction > 0)
+        max_load = [hvac_sizing_values.Heat_Load, hvac_sizing_values.Cool_Load_Tot].max
+        hvac_sizing_values.Heat_Load = max_load
+        hvac_sizing_values.Cool_Load_Sens *= max_load / hvac_sizing_values.Cool_Load_Tot
+        hvac_sizing_values.Cool_Load_Lat *= max_load / hvac_sizing_values.Cool_Load_Tot
+        hvac_sizing_values.Cool_Load_Tot = max_load
+
+        # Override Manual S oversize allowances:
+        hvac.OverSizeLimit = 1.0
+        hvac.OverSizeDelta = 0.0
+      end
+    end
+  end
+
   def self.get_duct_regain_factor(duct)
     # dse_Fregain values comes from MJ8 pg 204 and Walker (1998) "Technical background for default
     # values used for forced air systems in proposed ASHRAE Std. 152"
@@ -1964,16 +1986,21 @@ class HVACSizing
 
     heat_cap_rated = (heating_load / MathTools.biquadratic(@heat_setpoint, heating_db, coefficients)) / capacity_ratio
 
-    if totalCap_CurveValue.nil?
-      # Heat pump has no cooling, size based on heating
-      hvac_sizing_values.Heat_Capacity = heat_cap_rated
+    if totalCap_CurveValue.nil? # Heat pump has no cooling
+      if @hpxml.header.heat_pump_sizing_methodology == HPXML::HeatPumpSizingMaxLoad
+        # Size based on heating, taking into account reduced heat pump capacity at the design temperature
+        hvac_sizing_values.Heat_Capacity = heat_cap_rated
+      else
+        # Size equal to heating design load
+        hvac_sizing_values.Heat_Capacity = hvac_sizing_values.Heat_Load
+      end
     elsif heat_cap_rated < hvac_sizing_values.Cool_Capacity
       # Size based on cooling
       hvac_sizing_values.Heat_Capacity = hvac_sizing_values.Cool_Capacity
     else
       cfm_per_btuh = hvac_sizing_values.Cool_Airflow / hvac_sizing_values.Cool_Capacity
-      if @hpxml.header.use_max_load_for_heat_pumps
-        # Size based on heating
+      if @hpxml.header.heat_pump_sizing_methodology == HPXML::HeatPumpSizingMaxLoad
+        # Size based on heating, taking into account reduced heat pump capacity at the design temperature
         hvac_sizing_values.Cool_Capacity = heat_cap_rated
       else
         # Size based on cooling, but with ACCA oversizing allowances for heating